requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
google-generativeai==0.3.2
speechrecognition==3.10.0
pydub==0.25.1
gtts==2.4.0
webrtcvad==2.0.10
numpy==1.24.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-socketio==5.10.0
pyaudio==0.2.11 